criado tratamentos no relatorio e obrigatoriedade de filial e salario

+++ Gerenciador/src/service/FuncionarioService.java
@@ -1,6 +1,6 @@
 package service;
 
-import java.math.BigDecimal;
+
 import java.util.List;
 
 import javax.ejb.Stateless;
@@ -87,7 +87,7 @@ public class FuncionarioService extends GenericService<Funcionario>{
 		
 	}
 	
-	public List<Funcionario> listarFuncionarioValorSalarialSemFilial(BigDecimal salarioInicial, BigDecimal salarioFinal){
+	public List<Funcionario> listarFuncionarioValorSalarialSemFilial(Double salarioInicial, Double salarioFinal){
 		final CriteriaBuilder cBuilder = getEntityManager().getCriteriaBuilder();
 		final CriteriaQuery<Funcionario> cQuery = cBuilder.createQuery(Funcionario.class);
 		final Root<Funcionario> rootFuncionario = cQuery.from(Funcionario.class);
@@ -95,7 +95,7 @@ public class FuncionarioService extends GenericService<Funcionario>{
 		final Expression<String> expSalario = rootFuncionario.get("salario");
 		
 		cQuery.select(rootFuncionario);
-		cQuery.where(cBuilder.between(rootFuncionario.<BigDecimal>get("salario"),salarioInicial,salarioFinal));
+		cQuery.where(cBuilder.between(rootFuncionario.<Double>get("salario"),salarioInicial,salarioFinal));
 		cQuery.orderBy(cBuilder.desc(expSalario));
 		
 		
@@ -105,7 +105,8 @@ public class FuncionarioService extends GenericService<Funcionario>{
 		return resultado;	
 	}
 	
-	public List<Funcionario> listarFuncionarioValorSalarialComFilial(BigDecimal salarioInicial, BigDecimal salarioFinal, Long idFilial){
+	
+	public List<Funcionario> listarFuncionarioValorSalarialComFilial(Double salarioInicial, Double salarioFinal, Long idFilial){
 		final CriteriaBuilder cBuilder = getEntityManager().getCriteriaBuilder();
 		final CriteriaQuery<Funcionario> cQuery = cBuilder.createQuery(Funcionario.class);
 		final Root<Funcionario> rootFuncionario = cQuery.from(Funcionario.class);
@@ -113,7 +114,7 @@ public class FuncionarioService extends GenericService<Funcionario>{
 		final Expression<String> expSalario = rootFuncionario.get("salario");
 		
 		cQuery.select(rootFuncionario);
-		cQuery.where(cBuilder.and(cBuilder.equal(rootFuncionario.get("filial").get("id"),idFilial), cBuilder.between(rootFuncionario.<BigDecimal>get("salario"),salarioInicial,salarioFinal)));
+		cQuery.where(cBuilder.and(cBuilder.equal(rootFuncionario.get("filial").get("id"),idFilial), cBuilder.between(rootFuncionario.<Double>get("salario"),salarioInicial,salarioFinal)));
 		cQuery.orderBy(cBuilder.desc(expSalario));
